requirements.txt
anyio==4.10.0
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
djangorestframework==3.16.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
python-telegram-bot==22.4
requests==2.32.5
sniffio==1.3.1
sqlparse==0.5.3
urllib3==2.5.0
